SU7_2.py
Write a Python script that prompts the user to enter two numbers and an operator (+, ‐, *, /) to
perform a mathemaƟcal operaƟon on those numbers. The program should check for a division by
zero error and print an error message if the user aƩempts to divide by zero. The program should also
use if‐elif‐else statements to perform the specified operaƟon and print the result.
Message examples
 The calculated result
 Error – division by zero
 Error – invalid operator