Create SU7_3.py
Write a Python script that prompts the user to enter a base and an exponent. The program should
use if statements to check whether the exponent is negaƟve. The program should use or, and, or not
operators to check if BOTH the base and exponent are not zero (since 0 to the power of 0 is not
defined). The program should then use the pow() funcƟon from the math module to perform
exponenƟaƟon of the base to the exponent.
Message examples
 The calculated result
 Error – exponent cannot be negaƟve
 Error – indefinite form (0˄0)