Create SU9_2.py
You need to provide a multiplication table (from 1 to 12) for the value entered by the user, 
utilising a for loop. 
Instructions
1. Ask the user to enter an integer value. The value entered by the user must greater than 
1 (one), therefore, apply input validation by using a while loop. Ask the user to enter a 
value until a valid entry is made. 
2. Display the multiplication table based on the user-input. Example of output if the user 
enters the value of 5:
3. Lastly, modify the script to determine the sum of the totals from the multiplication 
table, print the total. 
Make sure you supply headings to ensure that your input and output make sense.