Create SU9_3.py
a Python script that prompts the user for an email address, extract the name and domain from 
the address using a loop, and print the results. 
Instructions
1. Ask the user to enter an email address.
2. Loop through the characters in the email address. 
Hint: Remember that Python strings can be treated as arrays of characters, so you can use array 
indexing and slicing to extract substrings from the email address. 
E.g., emailString[index] with index meaning the position within the string, represented by 
an integer value.
3. Print out the name and domain as output. 
Notes: 
 The len() function in Python returns the length of a string (number of characters contained in the string). 
E.g., len(name) will return the value three (3) if the name contained “Sam”.
 Using block brackets with two integer values separated by a colon, e.g., [start:end], allows one to select 
parts of a string in Python, named a substring. 
E.g., emailString[:5] => will return the first characters from a string, for a total of five (5) characters. 
Or, emailString[6:] => will return the sixth character until the last character.